Values for highscore and collectables work now
collectables+level+inlevel are the collectables you achieved in this one game. collectables_total+level are the collectables you achieved in total in all games you played. Same with highscore, highscore+level+inlevel is the highscore in this one game -> for endscreen. highscore_+level is the highest score you reached